lib: Print better error for invalid baudrate
The way the serial crate in use is written it will accept any arbitrary
baudrate and try to configure the interface with it. This can fail if
the underlying hardware doesn't support the given baudrate.

Previously the code would simply output an `Error: Invalid argument`,
which leaves the user guessing which part of their input was invalid.

This patch adds an additional check on the error type returned by
configuring the serial device. If it returns an `InputError`, we assume
that in fact the baud rate was invalid (Because that is the only setting
we change) and print a better error message to the user.