Remove --bin option from cargo-espmonitor
This wasn't there prior to the clap-ification, and it gets overwritten
anyway, which could confuse people.

I don't love that this duplicates the contents of AppArgs in
CargoAppArgs, but oh well, what are you gonna do...

(I do recognize that cargo-espmonitor maybe should have a --bin option
in order to select the bin target for projects that have more than one.
But I'll revisit that later.)